Modelos nuevos y carga de tableros en la página principal

+++ app/src/main/java/com/example/protasks/models/User.kt
@@ -22,7 +22,7 @@ class User(@Expose
 
     @Expose
     @SerializedName("Id")
-    private val id: Int? = null
+    private val id: Long? = null
 
     override fun getName(): String? {
         return username
@@ -44,7 +44,7 @@ class User(@Expose
         return username
     }
 
-    override fun getId(): Int? {
+    override fun getId(): Long? {
         return id
     }
